# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ruian-backend
  description: "\n    ## 锐安后端 API 文档\n\n    本 API 提供以下功能：\n\n    * **认证管理**: 用户登录和身份验证\n\
    \    * **文件上传**: 支持本地存储和阿里云 OSS 两种方式上传 Excel 文件\n    * **生产信息查询**: 查询和管理生产订单信息\n\
    \    * **员工工作量查询**: 查询和管理员工工作量记录\n    * **数据核验**: 验证生产信息和员工工作量数据的一致性\n\n    ##\
    \ 认证说明\n\n    大部分接口需要 Bearer Token 认证。请先调用 `/api/auth/login` 接口获取 token，\n   \
    \ 然后在请求头中添加：`Authorization: Bearer <your_token>`\n\n    ## 错误响应\n\n    所有错误响应都遵循统一格式：\n\
    \    ```json\n    {\n        \"error\": true,\n        \"message\": \"错误消息\",\n\
    \        \"detail\": {}\n    }\n    ```\n    "
  version: 0.1.0
paths:
  /api/auth/login:
    post:
      tags:
      - auth
      summary: 用户登录
      description: 使用用户名和密码登录，获取访问令牌（JWT Token）。
      operationId: login_api_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
        required: true
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Login Api Auth Login Post
              example:
                access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                token_type: bearer
        '401':
          description: 认证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationErrorResponse'
        '422':
          description: 请求参数验证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorResponse'
...
  /api/worklog:
    get:
      tags:
      - worklog
      summary: 查询员工工作量列表
      description: 分页查询员工工作量，支持按订单号、日期范围、员工工号等条件过滤。需要 Bearer Token 认证。
      operationId: get_employee_worklog_api_worklog_get
      security:
      - HTTPBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: 页码，从 1 开始
          default: 1
          title: Page
        description: 页码，从 1 开始
        example: 1
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: 每页数量，最大 100
          default: 10
          title: Page Size
        description: 每页数量，最大 100
        example: 10
      - name: order_no
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 订单号（模糊匹配）
          title: Order No
        description: 订单号（模糊匹配）
        example: ORD20250101
      - name: start_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          description: 开始日期（基于 work_date），格式：YYYY-MM-DDTHH:MM:SS
          title: Start Date
        description: 开始日期（基于 work_date），格式：YYYY-MM-DDTHH:MM:SS
        example: '2025-01-01T00:00:00'
      - name: end_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          description: 结束日期（基于 work_date），格式：YYYY-MM-DDTHH:MM:SS
          title: End Date
        description: 结束日期（基于 work_date），格式：YYYY-MM-DDTHH:MM:SS
        example: '2025-01-31T23:59:59'
      - name: employee_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 员工工号（精确匹配）
          title: Employee Id
        description: 员工工号（精确匹配）
        example: EMP001
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeWorklogListResponse'
              example:
                total: 100
                page: 1
                page_size: 10
                items:
                - id: 1
                  order_no: ORD20250101001
                  model: MODEL-A
                  brand_no: BRAND-001
                  employee_id: EMP001
                  employee_name: 张三
                  job_type: 组装
                  quantity: 50
                  performance_factor: 1.2
                  performance_amount: 1200.0
                  work_date: '2025-01-01T00:00:00Z'
                  upload_date: '2025-01-01T00:00:00Z'
                  validation_result: 正常
                  created_at: '2025-01-01T00:00:00Z'
                  updated_at: '2025-01-01T00:00:00Z'
        '401':
          description: 认证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationErrorResponse'
        '422':
          description: 请求参数验证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorResponse'
  /api/worklog/{order_no}:
    get:
      tags:
      - worklog
      summary: 按订单号查询员工工作量
      description: 根据订单号精确查询所有相关的员工工作量记录。需要 Bearer Token 认证。
      operationId: get_employee_worklog_by_order_no_api_worklog__order_no__get
      security:
      - HTTPBearer: []
      parameters:
      - name: order_no
        in: path
        required: true
        schema:
          type: string
          title: Order No
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmployeeWorklogResponse'
                title: Response Get Employee Worklog By Order No Api Worklog  Order
                  No  Get
              example:
              - id: 1
                order_no: ORD20250101001
                model: MODEL-A
                brand_no: BRAND-001
                employee_id: EMP001
                employee_name: 张三
                job_type: 组装
                quantity: 50
                performance_factor: 1.2
                performance_amount: 1200.0
                work_date: '2025-01-01T00:00:00Z'
                upload_date: '2025-01-01T00:00:00Z'
                validation_result: 正常
                created_at: '2025-01-01T00:00:00Z'
                updated_at: '2025-01-01T00:00:00Z'
        '401':
          description: 认证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationErrorResponse'
        '404':
          description: 订单号不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AuthenticationErrorResponse:
      properties:
        error:
          type: boolean
          title: Error
          description: 是否为错误响应
          default: true
        message:
          type: string
          title: Message
          description: 错误消息
        detail:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Detail
          description: 错误详情
      type: object
      required:
      - message
      title: AuthenticationErrorResponse
      description: 认证错误响应模型
      example:
        detail: {}
        error: true
        message: 认证失败
...
    EmployeeWorklogListResponse:
      properties:
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
        items:
          items:
            $ref: '#/components/schemas/EmployeeWorklogResponse'
          type: array
          title: Items
      type: object
      required:
      - total
      - page
      - page_size
      - items
      title: EmployeeWorklogListResponse
      description: 员工工作量列表响应模型
    EmployeeWorklogResponse:
      properties:
        id:
          type: integer
          title: Id
          description: 记录 ID
        order_no:
          type: string
          title: Order No
          description: 订单号
        model:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
          description: 型号
        brand_no:
          anyOf:
          - type: string
          - type: 'null'
          title: Brand No
          description: 品牌编号
        employee_id:
          type: string
          title: Employee Id
          description: 员工工号
        employee_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Employee Name
          description: 员工姓名
        job_type:
          type: string
          title: Job Type
          description: 作业类型
        quantity:
          type: integer
          title: Quantity
          description: 数量
        performance_factor:
          type: number
          title: Performance Factor
          description: 绩效系数
        performance_amount:
          type: number
          title: Performance Amount
          description: 绩效金额
        work_date:
          type: string
          format: date-time
          title: Work Date
          description: 工作日期
        upload_date:
          type: string
          format: date-time
          title: Upload Date
          description: 上传日期
        validation_result:
          type: string
          title: Validation Result
          description: 核验结果
        created_at:
          type: string
          format: date-time
          title: Created At
          description: 创建时间
        updated_at:
          type: string
          format: date-time
          title: Updated At
          description: 更新时间
      type: object
      required:
      - id
      - order_no
      - employee_id
      - job_type
      - quantity
      - performance_factor
      - performance_amount
      - work_date
      - upload_date
      - validation_result
      - created_at
      - updated_at
      title: EmployeeWorklogResponse
      description: 员工工作量响应模型
      example:
        brand_no: BRAND-001
        created_at: '2025-01-01T00:00:00Z'
        employee_id: EMP001
        employee_name: 张三
        id: 1
        job_type: 组装
        model: MODEL-A
        order_no: ORD20250101001
        performance_amount: 1200.0
        performance_factor: 1.2
        quantity: 50
        updated_at: '2025-01-01T00:00:00Z'
        upload_date: '2025-01-01T00:00:00Z'
        validation_result: 正常
        work_date: '2025-01-01T00:00:00Z'
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    InternalServerErrorResponse:
      properties:
        error:
          type: boolean
          title: Error
          description: 是否为错误响应
          default: true
        message:
          type: string
          title: Message
          description: 错误消息
        detail:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Detail
          description: 错误详情
      type: object
      required:
      - message
      title: InternalServerErrorResponse
      description: 服务器内部错误响应模型
      example:
        detail:
          message: Connection timeout
          type: DatabaseError
        error: true
        message: 服务器内部错误
    NotFoundErrorResponse:
      properties:
        error:
          type: boolean
          title: Error
          description: 是否为错误响应
          default: true
        message:
          type: string
          title: Message
          description: 错误消息
        detail:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Detail
          description: 错误详情
      type: object
      required:
      - message
      title: NotFoundErrorResponse
      description: 资源未找到错误响应模型
      example:
        detail:
          id: '12345'
          resource: order
        error: true
        message: 资源未找到
...
    UserLogin:
      properties:
        username:
          type: string
          maxLength: 50
          minLength: 1
          title: Username
          description: 用户名
        password:
          type: string
          minLength: 6
          title: Password
          description: 密码
      type: object
      required:
      - username
      - password
      title: UserLogin
      description: 用户登录模型
      example:
        password: password123
        username: admin
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    ValidationErrorResponse:
      properties:
        error:
          type: boolean
          title: Error
          description: 是否为错误响应
          default: true
        message:
          type: string
          title: Message
          description: 错误消息
        detail:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Detail
          description: 错误详情
      type: object
      required:
      - message
      title: ValidationErrorResponse
      description: 验证错误响应模型
      example:
        detail:
          field: order_no
          reason: 订单号不能为空
        error: true
        message: 数据验证失败
...
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 输入 JWT token，格式：Bearer <token>
...